:wrench: style(part3): Stops inserting a final newline
When the file is saved, the formatting is applied. When autosave is enabled, the code is frequently formatted. Additionally, the formatter is configured to add a final newline to the file. In new files, this behavior is highly inconvenient because it keeps moving the cursor to the next line WHILE I TYPE, messing my code and forcing me to stop and fix it. To stop it, I disabled this option in the editorconfig file.